refactor: move ThemeMappings to /src/configs
