--- src/pages/Settings/GroupThemeColorPage/ThemeMappings.ts
@@ -1,102 +1,0 @@
-import type { GroupTheme } from "@/types/Group"
-
-export const themeColorMappings: { [k: string]: GroupTheme } = {
-    defaultTheme: {
-        primaryColor: "#F6BC0D",
-        secondaryColor: "#BE392C",
-        tertiaryColor: "#FFFFFF",
-        backgroundColor: "#F5F4F4",
-        cardBackgroundColor: "#D9D9D9",
-        cardItemColor: "#F3F3F3",
-        fontColorV1: "#FFFFFF",
-        fontColorV2: "#191919",
-        fontColorV3: "#8A8A8A",
-        fontColorV4: "#7D7EAE",
-        fontColorV5: "#F5F5F5",
-        fontColorV6: "#4E4E4E",
-        fontDisabledColor: "#6F6F6F",
-        formsColor: "#E0E0E0",
-        skills1Color: "#BE392C",
-        waveColor: "#9294CC",
-        buttonYellowHoverColor: "#D3A61E",
-        buttonHoverColor: "#FFFFFF",
-        alertColor: "#CC615B",
-        successColor: "#35BD00",
-        wrongInvalidColor: "#B33B3B",
-        rankColor: "#6E70AF",
-    },
-    themeId1: {
-      primaryColor: "#0091B9",
-      secondaryColor: "#934588",
-      tertiaryColor: "#FFFFFF",
-      backgroundColor: "#F5F4F4",
-      cardBackgroundColor: "#D9D9D9",
-      cardItemColor: "#F3F3F3",
-      fontColorV1: "#FFFFFF",
-      fontColorV2: "#191919",
-      fontColorV3: "#8A8A8A",
-      fontColorV4: "#7D7EAE",
-      fontColorV5: "#F5F5F5",
-      fontColorV6: "#4E4E4E",
-      fontDisabledColor: "#6F6F6F",
-      formsColor: "#E0E0E0",
-      skills1Color: "#934588",
-      waveColor: "#9294CC",
-      buttonYellowHoverColor: "#D3A61E",
-      buttonHoverColor: "#FFFFFF",
-      alertColor: "#CC615B",
-      successColor: "#35BD00",
-      wrongInvalidColor: "#B33B3B",
-      rankColor: "#6E70AF",
-    },
-    themeId2: {
-      primaryColor: "#02366E",
-      secondaryColor: "#FA7511",
-      tertiaryColor: "#FFFFFF",
-      backgroundColor: "#F5F4F4",
-      cardBackgroundColor: "#D9D9D9",
-      cardItemColor: "#F3F3F3",
-      fontColorV1: "#FFFFFF",
-      fontColorV2: "#191919",
-      fontColorV3: "#8A8A8A",
-      fontColorV4: "#7D7EAE",
-      fontColorV5: "#F5F5F5",
-      fontColorV6: "#4E4E4E",
-      fontDisabledColor: "#6F6F6F",
-      formsColor: "#E0E0E0",
-      skills1Color: "#934588",
-      waveColor: "#9294CC",
-      buttonYellowHoverColor: "#D3A61E",
-      buttonHoverColor: "#FFFFFF",
-      alertColor: "#CC615B",
-      successColor: "#35BD00",
-      wrongInvalidColor: "#B33B3B",
-      rankColor: "#6E70AF",
-    },
-    themeId3: {
-      primaryColor: "#D2123B",
-      secondaryColor: "#D2123B",
-      tertiaryColor: "#FFFFFF",
-      backgroundColor: "#F5F4F4",
-      cardBackgroundColor: "#D9D9D9",
-      cardItemColor: "#F3F3F3",
-      fontColorV1: "#FFFFFF",
-      fontColorV2: "#191919",
-      fontColorV3: "#8A8A8A",
-      fontColorV4: "#7D7EAE",
-      fontColorV5: "#F5F5F5",
-      fontColorV6: "#4E4E4E",
-      fontDisabledColor: "#6F6F6F",
-      formsColor: "#E0E0E0",
-      skills1Color: "#934588",
-      waveColor: "#9294CC",
-      buttonYellowHoverColor: "#D3A61E",
-      buttonHoverColor: "#FFFFFF",
-      alertColor: "#CC615B",
-      successColor: "#35BD00",
-      wrongInvalidColor: "#B33B3B",
-      rankColor: "#6E70AF",
-    },
-};
-
-export default themeColorMappings;
